Have an explicit order of searching for stubs when parsing
Summary:
`typeshed/stdlib/3.4/enum.pyi` and `typeshed/third-party/3/enum.pyi` both exist, which means that our existing parsing
logic will parse both of these modules and create modules for them - since we had a deterministic parsing order, we were
able to ignore most of the issues with these conflicts, but the existence of both sources means that we could have a race when
building the environment in parallel. Add an explicit order of preference for keeping stubs (source root > search path > stdlib > third_party) in order to prevent this.

Reviewed By: dkgi

Differential Revision: D9125662

fbshipit-source-id: 005d620f4f99fc38c93622839bb341f0d57294d4