initial implementation of pyre-monitor
Summary:
This diff sets up the piping for the pyre-monitor and does the simple "kill pyre when a local configuration gets updated" logic.
Uses existing file locking infrastructure to ensure only one exists, and a double forking trick to ensure that its lifetime is separate from
that of its caller.

The timeout of 3600s on the pywatchman client means that the monitor will die after 1 hour of no updates from watchman.

Reviewed By: dkgi

Differential Revision: D9092438

fbshipit-source-id: b1134dbe78c923952cfe8261bb068aa609a4beaf