handle unix errors when building initial symlink map
Summary:
We were not handling unix errors when building the initial symlink map, which meant that the watchman client might fail to initialize if a broken symlink exists in the filesystem.

Changes the API of and exposes the build_symlink_map function to allow testing, since we don't have mocks.

Reviewed By: dark

Differential Revision: D9226996

fbshipit-source-id: a577a657c2a1fda6e4d9e670be7f5b1b608d9dfe