simplify metaclass implementation
Summary:
The existing metaclass lookup was hidden within the attributes and was hard to understand due
to the successor fold no longer having the metaclasses registered as superclasses of the current class. Pulls the metaclass
lookup out and always adds the entire metaclass chain via superclasses. This change makes it so that
we always look for the superclasses of the metaclass for class attributes, including for type.

Reviewed By: dkgi

Differential Revision: D9254304

fbshipit-source-id: 030d948d7119d765ce2e71f47b46339b52e95057