simplify default type order building
Summary:
Simplifies the type order building to only take a list of inheritance chain lists, and ensures everything's inserted into the
type order.

So, if you have `[A, B, C]`, the new code will ensure that `A`, `B` and `C` are inserted into the type order and add the `A -> B`, `B -> C` edges.

No behavioral change for now.

Reviewed By: dkgi

Differential Revision: D9258590

fbshipit-source-id: 06f1169ef51a23ab59bf10c837fd19a1e4a23907