Skip typeshed directories we are not interested in
Summary:
Ignore some directories of the typeshed repository that we know are
not interested in. Specifically, the "tests" directory, and everything
that starts with a dot (e.g., ".git", ".hg", etc...).

This is mostly heplful for folks that want to try a new typeshed
version, hence might use `--typeshed` to point to a checked-out
typeshed repository.

Reviewed By: sinancepel

Differential Revision: D9273875

fbshipit-source-id: d6c2714982e43fde369aeca2e6789d9e32057933