remove all calls to Annotated.resolve and Resolution.resolve in AnalysisTypeCheck
Summary:
Most changes are straightforward: Use forward_expression rather than Annotated.resolve in order to also
propagate errors, and do not ignore those errors whenever possible.

The one tricky bit is `initial` - it used `Annotated.resolve` but was declared before forward_expression and forward_statement. It needed to
be mutually recursive with `forward_expression` for this reason. Make it so.

Reviewed By: dkgi

Differential Revision: D9273377

fbshipit-source-id: 15f7ea6f3f58f662f79621fbe67233b0176a98ff