Create shared source directory per source root
Summary: Keeping track of when we actually are going off of a local configuration file (rather than searching for config options in a project root config or via command line args), and where that local config lives. We'll use this info when creating a shared source directory per project in scratch.

Reviewed By: sinancepel

Differential Revision: D9254657

fbshipit-source-id: ea4bbe02f6ba3fb4c16eaf23c78b892a4b7767e5