always return annotations when registering class definitions except for builtins
Summary:
For context: The server was not picking up on updates to class types due to the classes preexisting. This diff makes it so
that classes that already exist get repopulated. It's not easy to demonstrate the case that fails on this diff, but it has to do with
the order of superclasses getting flipped when the server gets updated, and an integration test will be added once deduplicate is changed to not reverse
the list in a future diff.

Reviewed By: dkgi

Differential Revision: D9259240

fbshipit-source-id: 9014ac69350ae237e9426aae73432d0e1263a060