Add test coverage for ClassVar unwrapping in static methods
Summary:
I may have missed a diff fixing this, but up until today the behavior on master for this test was to throw:

```
Incompatible return type [7]: Expected `str` but got `ClassVar[str]`.
```

May as well add in the test I was going to use to debug this... just in case this wasn't explicitly fixed with its own testing somewhere else (couldn't find).

Reviewed By: sinancepel

Differential Revision: D9331633

fbshipit-source-id: cd27b164f34a530f5981b2b3c22ac7a56bc81712