always reconstruct shared source directory
Summary:
`pyre restart` wasn't properly reconstructing shared source directories because the contents of said directories could change,
even though the analyzed directories stay consistent. If we were to do this properly, maintaining the cache and ensuring consistency would be a lot of work, so
defer on the staleness checking until we can make that happen.

Reviewed By: dkgi

Differential Revision: D9349004

fbshipit-source-id: 6ff44b4f7cd06daf422a9b0d54fb98a52a856aaf