Introduce parameter map type
Summary:
help wanted to make this better :(

I want to create a Map of parameters, so that I can start restructuring signature select to build up a map from params to args that have been matched to that param slot.

-----more context

The eventual goal here is to have a Map type going from `parameter` -> `argument list`, so when we do signature selection, I can create one "bucket" per parameter slot and dump all arguments that have been matched to that parameter into the bucket. These will only contain one argument unless the parameter is * or **. Second pass, we'll go through typechecking the buckets (see suggestion on D8622751)

Issues with this diff: renaming is hard... the record 'Callable' needs to be included into the non-record 'Callable', but a 'Parameter' module is defined in both and the names conflict. Ideally I'd be able to include `Record.Callable` without the `Parameter` module, or include it and rename the `Record.Callable.Parameter` module into something different in that scope to avoid the name clash.

I was also getting very confusing errors when attempting a workaround by just placing the Map in the Callable module itself (whtout trying to create a Parameter submodule)... I tried creating a `type parameter = type_t Record.Callable.Parameter.t` and generating `compare, eq, sexp`, etc. on it. However, when passing `compare` into Map.Make, the compiler thinks it is `type_t -> type_t -> bool` rather than `parameter -> parameter -> bool`. :( If this section is unclear, I'm happy to re-implement this and put it up so you can see what it looks like.

Reviewed By: dkgi

Differential Revision: D9346525

fbshipit-source-id: 79a169c5b2f98935040d7c20f3bd04eeda6d5fb3