only prepare shared analysis directory in start and check
Summary:
Preparing the shared analysis directory in non-start and check commands was causing a bad user experience (have to wait for errors)
as well as potential races with file updates and the server aiming to check something that's being removed. Simplify the
model to have only the commands that do the analysis (start and check) be able to modify the shared analysis directory.

Reviewed By: shannonzhu

Differential Revision: D9382072

fbshipit-source-id: 9197f005905dc5fa647b64b8b14d8e35c0c8868d