add builtins classes for None and typing.Optional
Summary:
We need `None` and `typing.Optional` to be classes in order for the attributes computation to take note of them being objects and picking up on
methods such as `__eq__` and `__class__`

Note that we don't need to have the same treatment for other special forms (Tuple, Type, Generic, Protocol, ClassVar). In the case of tuple and type, we already have a builtins.tuple/type class that we alias into, and "Protocol", "Generic" and "ClassVar" should never be passed as function arguments/appear in actual expressions.

Reviewed By: dkgi

Differential Revision: D9549567

fbshipit-source-id: 4a447760e174a4fd0fd96b57fa2cf1f5ec52ab6b