Search for a root consistently in File.handle
Summary:
Context: Currently, the File.handle abstraction is leaky, and allows users to specify any root they'd like and get
different handles for the same path based on their choice. Normalize this by passing in a configuration and doing the search
in a consistent order when creating.

This manifested itself as not picking up on external file updates if a directory in the search path was a subdirectory of the source root being
analyzed.

Reviewed By: dkgi

Differential Revision: D9561908

fbshipit-source-id: 5cd1ed381567911f54643bb2d181af6c02586920