use handle instead of relative paths for shared memory handles
Summary:
We were not picking up on changes to stubs under the local root on the server side due to always assuming that
changed files were under the local root, rather than somewhere in the search path. Modify the way we compute paths
for sources to be more consistent and have the server pick up on stub updates by this process.

Reviewed By: dkgi

Differential Revision: D9565796

fbshipit-source-id: 5da8f1bd5d82cca833395805d08d6e7521b4d57b