Fix flush trigger for statistics
Summary:
The previous code was using the number of categories to determine when
to flush the statistics map. The number of categories is unlikely to
exceed 500 (the threshold size).

Fix the code to follow the original intent - that is, flush the
statistics when the number of samples (across all categories) is
greater than the threshold.

Reviewed By: dkgi

Differential Revision: D9651711

fbshipit-source-id: 073dafcd81c89e2ed451d9976ce5e3b3b6c9dd40