Lookup method definitions
Summary:
Resolve types of class instances to be able to lookup method calls.

When we need to resolve an approximate definition entry, the default
that is applicable for annotations (i.e. take the value of the longest
access) is not appropriate: the lookup should fail instead. Add a flag
to control this behaviour.

Reviewed By: sinancepel

Differential Revision: D9253506

fbshipit-source-id: b2d6d97b1b71dbb9c1a441cd732a8e417bb2d8e6