Fixing do_not_check
Summary:
When trying to get xplat set up, I ran into some weird behavior with how do_not_check was being handled.
I discovered that there were basically three different interpretations of the paths in the do_not_check array, all of which were incompatible in the circumstance of using a local config.
* paths in the main configuration are put in verbatim, implying that they are relative to the main root.
* paths in the local configuration are put in as full absolute paths, implying that they are absolute paths
* when matching errors against them, they are treated as if they are relative to the original working directory by doing a fnmatch to a path made relative to the original working directory.

This meant that do_not_checks in the local config were never matched and do_not_matches in the main config needed to be written relative to the original working directory.

This diff makes all of the paths absolute and then matches off of that.

Reviewed By: sinancepel

Differential Revision: D9826838

fbshipit-source-id: 4483f99c602e6dd691bf00c07319e1986f0364da