simplify testing of stop command
Summary:
We currently have a graceful flag that doesn't exit if it is set in the `run_stop` function. Now that start and stop are separated,
we can split up the exiting from the computation of the exit code, and just use that function in tests instead of having an extra flag.

No behavioral change.

Reviewed By: dark

Differential Revision: D10035769

fbshipit-source-id: 3a0dd417e73e949e2a9f476f6332709319fef373