suppress OrderedDict/dict mismatches
Summary:
The big issue that crops up is named tuples. Their `asdict` function is annotated in typeshed as returning `OrderedDict[str, Any]`.
Unfortunately, it's hard to represent this well in the typeCheckTests, as the test would just show
that we indeed raise a error when you return a OrderedDict containing Any when we expected a dict.

Reviewed By: dark

Differential Revision: D10088914

fbshipit-source-id: 23312ff789db4f33b38af5264a3be0e7b6a8c80f