move exception handling in watchmanCommand inside `listen_for_changed_files`
Summary:
`listen_for_changed_files` is the main loop of the watchman command that gets called in both daemon and terminal mode. Any exceptions that happen earlier will be surfaced to the command line directly and don't need to be logged.

The existing Scheduler.run_process was only wrapping exceptions for the process that spawned the daemon, which is probably not what we want.

Reviewed By: dark

Differential Revision: D10148549

fbshipit-source-id: 846c3c3b8b7af05ff3b2328420ac2708a7b940fd