Do not suppress incompatible variable errors due to Any
Summary:
Don't silence incompatible variable errors because the type being assigned is Any. This can cause confusion around what the type of the variable is moving forward.

T30528615 was reported because an assignment was being made to a variable with an explicitly declared type which should have changed the type of that variable. However, because pyre wasn't able to infer the type of the assignment, it silently swallowed this and then complained later when the variable was used, as though the variable still had the former type.

This is super confusing, and I believe warrants reporting when Pyre decides not to change the annotation of a previously explicitly annotated type even when the R.H. side is Any (we currently warn if it any supertype of the previous annotation, except Any).

Minified example:
```
def expects_str(x: str) -> None:
    pass

def foo(x: int, y: Any) -> None:
    x = y
    expects_str(x)
```
^ This could get a lot more confusing if `y` came from somewhere else and looks like it has type `str`, but Pyre wasn't able to deduce this. See the linked task for a specific example.

Reviewed By: sinancepel

Differential Revision: D10095302

fbshipit-source-id: e304452a7d6b44c6630e443172aed49dbeb17a1b