Prioritize existing modules for given qualifiers
Summary:
If a user edited both file.py and file.pyi, the Pyre server's behavior on which one to update was nondeterministic.

This diff changes the behavior to always prioritize existing modules - a server, when being started, will prefer stubs over python files when both exist for the same qualifier, and this change will make it so that the original preference is respected by making sure that the qualifier being updated matches the existing file.

Reviewed By: dkgi

Differential Revision: D10153111

fbshipit-source-id: e854b14a84f30856b283332931cbbc8bfd3cb49f