Change Watchman query to be compatible with pre-5.0 versions
Summary:
The previous query was using features available only in the master
version of watchman. Watchman version 4.9.0, the latest released
version as of now, would reject this query at runtime, effectively
disabling watchman support in all cases.

Fix the problem by using an equivalent query that works on 4.9.0 as
well.

This closes #93 and closes #96.

Reviewed By: sinancepel

Differential Revision: D10225011

fbshipit-source-id: 032a6da913625f94fad08abc36b87235b11b0e3e