Store postcondition for nodes in shared memory
Summary:
Evaluating performance difference on this.

Decided to give this a shot rather than a more hacky (but less duplicated storage) approach, which would be to store the post condition of the last statement as some special key, and then when accessing the postconditions in astExport fixpoint, pass in the previous statement in each iteration to store the value of the lookup under each time, looking up the "fake" end statement in the last case.

That's gross, but currently we do not store the final resolution of each define anywhere else and so this wouldn't be accessible after typecheck is over any other way. If adding extra storage to the existing shared memory per statement doesn't have any meaningful perf impact, doing the extra storage might be preferable (and also make it more obvious that we're getting a precondition).

Reviewed By: sinancepel

Differential Revision: D10250799

fbshipit-source-id: 5dee9de0b8cb2aa0094b0749ef203cc5197bc98d