be open to unknown fields when decoding LSP json
Summary:
The language server protocol capabilities are unfortunately a moving target. New fields get added over time. In order
to not crash and burn when we get a new capability that Pyre doesn't understand, use the `strict = false` option of yojson.

All this does is ignore unknown fields, and we only add the strict = false fields to capability jsons.

Reviewed By: shannonzhu

Differential Revision: D10373363

fbshipit-source-id: d18add9619ba0e96445586b8e1a7d121d2b0708b