don't attempt to load a saved state when PYRE_BINARY is set
Summary:
Pyre believes the version in the configuration that gets passed in via -expected-binary-version when loading a saved state, but this expected binary version will be a lie when PYRE_BINARY is set, and the shared memory structure can change in general.

Therefore, do not try to load from saved states from a watchman-supported project when the actual binary version disagrees with the configuration's expected version. I decided to do this here instead of the python client because users still had the capability to load using pyre --saved-state-project name start, which we also want to disallow.

Reviewed By: shannonzhu

Differential Revision: D10378002

fbshipit-source-id: 80cea3ae086c5823f2701fa36bfc28c9b33e32f8