Warn when code is unreachable due to impossible isinstance check
Summary: Instead of silently not checking all code under an unreachable branch, we should warn that this is happening by when we detect the impossible isinstance check and make the decision to skip typechecking. Ideally this can avoid confusion if anything in the unreachable body looks like a type error but is not throwing.

Reviewed By: sinancepel

Differential Revision: D10432156

fbshipit-source-id: 0795c10e30a3d1cc573671a5ea4c02632e2b606d