Type check functions defined after a try/except block
Summary:
Due to missing logic in widen and join, the list of nested defines
(including the defines "nested" in the toplevel) was not always
correctly propagated in the fixpoint across nodes with multiple
predecessors. This was causing pyre to miss those definitions, and not
typecheck them.

Having a try/except/finally block at toplevel is quite common, as it
is widely used to handle ImportError exceptions.

Reviewed By: dkgi

Differential Revision: D10455568

fbshipit-source-id: f15aeb5ca8db6b872a50b8faee431a52fe001403