Do not use sequential runner to run tests
Summary:
The sequential runner suffers of a problem when tests are run: if a
test quits the process, further tests in the same module will not be
run, silently. This was the case for a few (broken) server tests, that
are also fixed in this diff. Specifically, avoid exiting on invalid
requests in Request.process: it's untestable and not robust.

Tests should be able to run in parallel, but at extreme parallelism
levels (implied by `dune -j auto`) the host might not have enough
resources to run them all. Hence, limit the number of shards used by
OUnit to run tests to 1 ("shards" is just a fancy term used by OUNit
to describe how many parallel workers should be used). This solution
keeps the same runtime of the previous version, but it avoids the
shortcomings of the sequential runner previously explained.

Reviewed By: dkgi

Differential Revision: D10493424

fbshipit-source-id: b4daeff64f4b0721e6848aa39b69045fe94dfff3