Simplify choice between overload stubs and implementations
Summary:
This can be simplified further since there is only one implementation associated with each overload. Need to untangle this from __getitem__ work and will stack it next.

Essentially, we know that as long as overload_stubs is not empty we should be selecting signatures from it. We don't need to go through and choose which to keep from the list of implementations.

Reviewed By: mrkmndz

Differential Revision: D10859418

fbshipit-source-id: 59ce0348bcac43dabe598af46319b31f2bac8a96