Do not throw contradiction on isinstance checks
Summary:
If we know that X is type Foo, isinstance(x, Bar) when Foo and Bar have no relation is not necessarily false.

X could have type Baz, which implements both Foo and Bar.

It makes sense that not isinstance() can create impossible situations, but just because A is not a subclass of B does not mean that something that is a subclass of A cannot also be a subclass of B.

Reviewed By: dark

Differential Revision: D12462259

fbshipit-source-id: edc7c46a662e0915b6d897f084431969c7ac29c2