Consider some assert statements as "synthetic" for lookup purposes
Summary:
Pyre creates internal assertions for the two branches of an if
statement, asserting either that the test expression is true or false.

When the test is against a simple value, like when checking whether a
container is "truthy" or not, this might cause the type of the
container to be treated as "bool", depending on the specific cfg and
iterations of the fixpoint; this causes the lookup to display a
misleading hover annotation.

Since this is clearly wrong, mark the negated expression we assert on
(that is causing pyre to treat the container as `bool`) as
"synthetic", by setting its location to `any`. This will make the
Lookup module ignore its type.

Reviewed By: dkgi

Differential Revision: D10861860

fbshipit-source-id: fe11afbc84dd4304f1404071b3f3c457598e0e8c