When server fails, include the error message in the log
Summary:
In debugging other issues with the server, I occasionally run into the message that there's an error when waiting for the server (on incremental run) with the suggestion to run `pyre restart`. Running `pyre restart` only outputs the same error again. P60264063 P60269421

This is frustrating, but it's hard to reproduce on demand. Hoping that adding this logging will make issues easier to debug.

Reviewed By: dark

Differential Revision: D12873505

fbshipit-source-id: 9984bfd27634229f7e5a1a59b0aff6fb4a401353