Typecheck generated files unless opted out
Summary:
We've gotten several requests to stop ignoring generated files, which can have important type information and are especially essential for nodeapi to strictify and expose type info to distillery.

We were previously ignoring generated files in our initial push to onboard the first project or two, after running into some monster thrift files that tanked our typechecking time. See test plan for perf analysis of turning this on nearly a year later. Generated files can still be ignored if a `# pyre-do-not-check` comment is generated in a line at the top of the file.

Reviewed By: sinancepel

Differential Revision: D12889201

fbshipit-source-id: 9008a05f8d661cc11900964e8997196c42e76305