Default to implementation signature if no overloads are perfect match
Summary:
Turns out we're now getting a lot more coverage of objects returned from NodeApi, but sometimes attributes of those are unknown and functions expecting some particular type. Node uses a lot of overload function stubs, and our old logic used to just choose a random one if they all matched the same, meaning we are typing the return value incorrectly and causing a lot of bad type errors to show up. If none of the overloads match and an implementation signature exists, throw the error based on the implementation signature and get the return value from that, too (it should cover the possible types of the overloads).

This gets rid of most of the fixmes that were added here: D12884114

Reviewed By: sinancepel

Differential Revision: D12898248

fbshipit-source-id: b56b48b898bde88624b5366d9a0cbd330e27d5f5