Always log performance of long-running generated files
Summary: Realized that our current logging samples every 100 randomly, and we look at the longest-running files from this sample pool when visualizing the files that take the longest to typecheck. It seems important to have reliable data on the longest running generated files, since those have a greater potential to get ugly and need ignoring or cleanup.

Reviewed By: sinancepel

Differential Revision: D12914013

fbshipit-source-id: a1e9ea467ada2602612f2dd7bcd92364300746b8