eliminate anonymous parameters from callables
Summary:
Unnamed parameters do not exist in Python. To test, you can try:

```
f = lambda x: x
a = f(x=2)
```

This example hopefully motivates that it is at least sensible to consider removing anonymous parameters from the callable type.

We hope to merge the currently ad-hoc handling of callable signatures and that of defines in the future. Simplifying the callable type to
mirror that of defines directly will make this more tractable. The tradeoff, of course, is that the printed callable syntax is uglier, but it should
be equivalent.

Reviewed By: dkgi

Differential Revision: D12945788

fbshipit-source-id: 81807713383a5cec3855d0610c2df3512d15fc2d