generalize less_or_equal for callables
Summary: We used to only allow comparing anonymous callables - it makes sense to also allow comparing anonymous callables to named ones, which will be a common fashion in which functions get passed in in practce. Extend the behavioral subtyping check to ignore the kind of callable except when both functions are named, and add a bunch of regular callable tests just to demonstate existing behavior.

Reviewed By: shannonzhu

Differential Revision: D12966492

fbshipit-source-id: d37439d03fe58cabf382b319ffc75060ceadeb1f