Don't use the word 'positional' when we don't distinguish arg types
Summary:
We don't distinguish between missing positional arguments and missing keyword arguments. Sometimes, parameters are only able to be supplied by keyword (any parameter that follows a `*arg`, for instance). If those are missing, we still blanket print text in the error saying a "positional argument" is missing.

If we don't distinguish, then let's keep the error simple and not add confusion.

Ex. of a case where this happened: T35007681

Reviewed By: sinancepel

Differential Revision: D12990253

fbshipit-source-id: 8dd846246577a93e8f9a48915f2231a8503f20ef