Make type checking state go to Bottom on `assert False`
Summary:
Saw code in the wild of the form
```
if Cond:
  kind = "One"
elif OtherCond:
  kind = "Two"
else:
  assert False, "Pattern match above was exhaustive."
send_message(kind)
```

This change makes it so that we set bottom in the `else` case and respect the type of `kind`.

I had to change our tests (we used if True all over the place, which means `assert False` in the else) to use
an opaque `condition()` to now swallow errors.

Reviewed By: shannonzhu

Differential Revision: D13003137

fbshipit-source-id: 18ec662c3367b0b18474e00326328754f3923d37