kill Annotated.resolve
Summary:
Replaces the implementation of Resolution.resolve with `forward_expression`, eliminating the need for Annotated.resolve.

Due to a circular dependency adventure between creating a resolution and referencing typechecking state in a resolution, we need to break down the typechecking state into
resolution/non-resolution bits, most of the diff is ramifications of that fact.

Note that the tests demonstrate previously buggy ternary logic being fixed now.

Reviewed By: dkgi

Differential Revision: D13026808

fbshipit-source-id: cdec080d8efb02a68b94d9b8d7d7de8f7d9b4bd8