Relativize source directory paths when reading from configuration
Summary:
When specifying source directories, we are reading the paths as relative to the project root. This isn't very intuitive, because if you're in directory `a/b/c` and your `.pyre_configuration.local` reads `source_directories: ["."]`, this will mean pyre will run not on `a/b/c` but on whatever directory your `.pyre_configuration` lives in.

This didn't come up much because fbcode relies on targets (which are specified relative to fbcode, the project root) and distillery has a single source directory specified in `instagram_server` to be `distillery`.

Reviewed By: sinancepel

Differential Revision: D13108784

fbshipit-source-id: 340cf11293274e8e6d973b669283080ac8605e30