Rewrite call to TypedDictionary constructor to type alias
Summary:
As alluded to in the preceding diff, here we rewrite a call to the mypy_extensions type constructor into a normal type alias, which then will get picked up as normal and used, using normal alias semantics.

In order to support all of mypy's declaration interfaces we now just need:
* kwargs version `Movie = TypedDict('Movie', name=str, year=int)` [simple addition to this one]
* class version `class Movie(TypedDict):\n name: str\n  year: int` [similar rewrite to this one, little more complicated
* arbitrary depth inheritance of classes [ we are not currently planning on prioritizing this, possibly down the line.  It presents a lot of challenges we don't necessarily need right now]

Structural subtyping and instance construction also in the works.

Reviewed By: sinancepel

Differential Revision: D12972816

fbshipit-source-id: 99a3368bf10a4fd4618226bbb4fca7af92c996bf