Hardcoding in a different version of the mypy_extensions stub
Summary:
We are using the `mypy_extensions.TypedDict[('<NAME>', <FIELDS>)]'` syntax but the `mypy_extensions` stubs have an entry for a callable by that name that is a factory for their types.  We are going to support that syntax by preprocessing that callable into an alias in the next diff, but in order for that syntax to not fail typechecking*, we need to add that to the stubs.

This is done in this diff by finding that stub during preprocessing and replacing the callable definition with an assignment with type `typing._SpecialForm`, which corresponds to our usage.

This now allows us to, after importing the stubs for `mypy_extensions` and `typing`, do a full, end to end test of the implemented functionality.

Doing this also exposed a bug where the typechecking logic was missing the fact that this was a type alias because `is_resolved` was hitting the name of the TypedDictionary when it shouldn't have (see type.ml)

(*) This doesn't actually fail typechecking right now because it gets interpreted as calling the `__getitem__` on the callable, which is of course wrong, but because we are currently swallowing all bad accesses on callables, it doesn't actually make an error.  This will soon change, so this diff is necessary.

Reviewed By: sinancepel

Differential Revision: D12972187

fbshipit-source-id: 30fc05b6eb475955535e5c2686d3e2bb0a9cf571