Introduce support for type variance
Summary:
This diff introduces support for type variance, as defined in:
https://www.python.org/dev/peps/pep-0483/#covariance-and-contravariance

Type variance is stored as part of Variable, and used inside
less_or_equal to determine subtype relations.

This diff aims at implementing the complete algorithm defined in PEP
483, including the Generic/Base and Manager/Employee example. It is
implemented as follows.

Given types A[T1, T2, ...] and B[T3, T4, ...], where all types are
concrete, the result of less_or_equal is computed with these steps:

1) If A === B, then the parameters are compared directly (T1 against
T3, T2 against T4, etc...). This comparison follows the variance rules
set forth by the PEP and defined in the class definition of A and B.

2) Otherwise, a multi-plane search is started. That is, the algorithm
tries and perform a step in all possible directions that would bring A
closer to B, and recurses on less_or_equal.

2.a) A step is taken in the types plane: all successors of A are
enumerated, and, for each of them, the parameters from A are
instantiated, and the search restarts recursively from the new
location.

2.b) If the step above fails, steps are taken in the parameters plane:
that is, we try and replace every parameter in A, one at a time
(e.g. from A[T1, T2] to A[T3, T2]), to see if that brings us closer to
B. This step is performed only if variance allows it (e.g. T3 replaces
T1 only if T1 < T3, taking into account the appropriate variance
definition). The search continues from the new location.

Reviewed By: sinancepel

Differential Revision: D13035416

fbshipit-source-id: 9dec1cd93276d8204d9d2a27eafb20f8f671b107